requirements.txt
opencv-python
openai
pyttsx3
python-dotenv
fastapi
uvicorn
requests
